requirements.txt
flask
wtforms
google.genai